+++ lib/socket.js
@@ -12,7 +12,6 @@ function Socket(options) {
   }
 
   EventEmitter.apply(this);
-  //var socket = new EventEmitter();
 
   options = options || {};
