+++ tests/test_intercept.js
@@ -8,7 +8,6 @@ var util    = require('util');
 var events  = require('events');
 var stream  = require('stream');
 var test    = require('tap').test;
-//var test = require('tape');
 var mikealRequest = require('request');
 var superagent = require('superagent');
 var needle  = require("needle");
@@ -3975,8 +3974,7 @@ test('match domain using intercept callback', function (t) {
     .reply(200, 'Match intercept');
 
   mikealRequest.get('http://www.interceptexample.com/cats', function(err, res, body) {
-    //t.type(err, 'null');
-    console.log(err);
+    t.type(err, 'null');
     t.equal(res.statusCode, 200);
     t.equal(body, 'Match intercept');
     t.end();
